+++ tests/testthat/test-small_helpers.R
@@ -3,34 +3,33 @@
 # on how much time they took.
 ###############################################################################
 
-dummy_df1 <- suppressMessages(dummy_data(10))
-dummy_df2 <- suppressMessages(dummy_data(10))
-dummy_df3 <- suppressMessages(dummy_data(10))
-
-
-test_that("Stack data fames", {
-    new_df1 <- set(dummy_df1, dummy_df2, dummy_df3)
-    new_df2 <- set(dummy_df1, dummy_df2, dummy_df3, compress = "factor")
-
-    expect_equal(nrow(new_df1), 30)
-    expect_equal(nrow(new_df2), 30)
-    expect_equal(class(new_df1[["education"]]), "character")
-    expect_equal(class(new_df2[["education"]]), "factor")
-})
-
-
-test_that("Stack data fames with id column", {
-    new_df <- set(dummy_df1, dummy_df2, dummy_df3, id = TRUE)
-
-    expect_equal(max(new_df[["ID"]]), 3)
-})
+dummy_df <- suppressMessages(dummy_data(100))
 
 
 test_that("Add new NA column range with helper function", {
-    added_df <- suppressMessages(dummy_df1 |>
+    added_df <- suppressMessages(dummy_df |>
           add_variable_range(status1:status3))
 
     expect_equal(utils::tail(names(added_df), 3), c("status1", "status2", "status3"))
+})
+
+
+test_that("Get length of integer variable", {
+    expect_equal(get_integer_length(1L), 1)
+    expect_equal(get_integer_length(10L), 2)
+    expect_equal(get_integer_length(100L), 3)
+})
+
+###############################################################################
+# Warning checks
+###############################################################################
+
+
+test_that("Warning in length of integer if numeric value provided", {
+    value <- 100.123
+    expect_message(get_integer_length(value), " ! WARNING: Variable is not an integer and will be floored. Decimal places won't count.")
+
+    expect_equal(get_integer_length(value), 3)
 })
 
 ###############################################################################
@@ -38,15 +37,20 @@ test_that("Add new NA column range with helper function", {
 ###############################################################################
 
 test_that("Adding NA variables to data frame aborts on wrong pattern", {
-    expect_message(added_df <- dummy_df1 |>
+    expect_message(added_df <- dummy_df |>
          add_variable_range(status1:age3), " X ERROR: Variable range has to be provided in the form 'var_name1:var_name10'.")
 })
 
 
 test_that("Adding NA variables aborts, if variables are already part of data frame", {
-    added_df <- suppressMessages(dummy_df1 |>
+    added_df <- suppressMessages(dummy_df |>
          add_variable_range(status1:status3))
 
     expect_message(added_df2 <- added_df |>
          add_variable_range(status1:status3), " X ERROR: Some variables are already part of the data frame: ")
+})
+
+
+test_that("Abort length of integer if character variable provided", {
+    expect_message(get_integer_length("Test"), " X ERROR: Only numeric values allowed.")
 })